requirements.txt
streamlit
streamlit-autorefresh
ultralytics
torch
torchvision
Pillow